Fix calibration bug where multiple curves couldn't share input data
